refactor: move `loadNuxtConfig` to `@nuxt/config`
