+++ packages/config/src/index.js
@@ -1,2 +1,3 @@
 export { defaultNuxtConfigFile, getDefaultNuxtConfig } from './config'
 export { getNuxtConfig } from './options'
+export { loadNuxtConfig } from './load'
